enemies can now be killed

+++ model/Weapon.hpp
@@ -3,10 +3,10 @@
 #include <FloorThing.hpp>
 #include <WorldCoordinates.h>
 
-constexpr int   WEAPON_HAND_DAMAGE = 5;
+constexpr float WEAPON_HAND_DAMAGE = 5.0f;
 constexpr float WEAPON_HAND_ATTACKRANGE = 30.0f;
 constexpr float WEAPON_HAND_COOLDOWN = 5.0f;
-constexpr int   WEAPON_AXE_DAMAGE = 20;
+constexpr float WEAPON_AXE_DAMAGE = 20.0f;
 constexpr float WEAPON_AXE_ATTACKRANGE = 15.0f;
 constexpr float WEAPON_AXE_COOLDOWN = 10.0f;
 
@@ -19,7 +19,7 @@ enum class WeaponType {
 class Weapon : public FloorThing {
 public:
     const WeaponType type;
-    const int damage;
+    const float damage;
     const float attackRange;
     const float coolDown;
     bool pickedUp;
